Move active_window to main package
Was moved to a sub-package because we had no way to run without showing
the GUI - now we can choose when to display the UI so this no longer
needs to be a separate package.
