+++ lib/active_window/src/logger.dart
