--- packages/active_window/.gitignore
@@ -1,6 +1,0 @@
-# Files and directories created by pub.
-.dart_tool/
-.packages
-
-# Conventional directory for build output.
-build/
