--- packages/active_window/README.md
@@ -1,1 +1,0 @@
-A simple command-line application.
