--- packages/active_window/pubspec.yaml
@@ -1,17 +1,0 @@
-name: active_window
-description: Toggle the suspend state of the active window
-version: 1.0.0
-publish_to: none
-
-environment:
-  sdk: ">=2.12.0 <3.0.0"
-
-dependencies:
-  args: ^2.2.0
-  hive: ^2.0.4
-  native_platform:
-    path: ../native_platform
-
-dev_dependencies:
-  lints: ^1.0.0
-  pedantic: ^1.11.1
