+++ pubspec.yaml
@@ -7,8 +7,6 @@ environment:
   sdk: ">=2.16.1 <3.0.0"
 
 dependencies:
-  active_window:
-    path: packages/active_window
   args: ^2.2.0
   desktop_integration:
     git:
@@ -19,6 +17,7 @@ dependencies:
   flutter_bloc: ^8.0.0
   flutter_markdown: ^0.6.9+1
   flutter_window_close: ^0.2.2
+  hive: ^2.0.4
   helpers:
     git:
       url: https://github.com/Merrit/helpers.git
